feat(agendamento): adiciona validacao de autorizacao por usuario nos agendamentos

+++ back-end/src/main/java/com/davi/restaurante/services/AgendamentoService.java
@@ -2,11 +2,15 @@ package com.davi.restaurante.services;
 
 import com.davi.restaurante.entity.AgendamentoEntity;
 import com.davi.restaurante.entity.MesaEntity;
+import com.davi.restaurante.entity.UsuarioEntity;
 import com.davi.restaurante.exceptions.AgendamentoException;
+import com.davi.restaurante.exceptions.RestauranteException;
+import com.davi.restaurante.exceptions.UsuarioException;
 import com.davi.restaurante.records.request.AgendamentoRecord;
 import com.davi.restaurante.records.response.AgendamentoResponseRecord;
 import com.davi.restaurante.repository.AgendamentoRepository;
 import com.davi.restaurante.repository.MesaRepository;
+import com.davi.restaurante.repository.UsuarioRepository;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -16,6 +20,7 @@ import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
 import java.util.List;
+import java.util.Objects;
 
 @Service
 public class AgendamentoService {
@@ -25,25 +30,37 @@ public class AgendamentoService {
     @Autowired
     private MesaRepository mesaRepository;
 
+    @Autowired
+    private UsuarioRepository userRepository;
+
     public AgendamentoService() {
     }
 
     public AgendamentoResponseRecord agendar(AgendamentoRecord record) {
         AgendamentoEntity agendamento = new AgendamentoEntity();
 
-        if (this.repository.existeConflitoNoAgendamento(record.data(), calcHoras(record.data(), record.duracao()), record.mesa().numero(), agendamento.getId()))
+        UsuarioEntity user = this.userRepository.findById(record.userId()).orElseThrow(() -> new UsuarioException("Usuário não encontrado", HttpStatus.NOT_FOUND));
+
+        MesaEntity mesa = this.mesaRepository.findByNumero(record.mesa().numero()).orElseThrow(() -> new RestauranteException("Mesa informada não encontrada", HttpStatus.NOT_FOUND));
+
+        if (this.repository.existeConflitoNoAgendamento(record.data(), calcHoras(record.data(), record.duracao()), mesa.getId(), null))
             throw new AgendamentoException("O tempo de agendamento gera conflito com os demais agendamentos", HttpStatus.CONFLICT);
 
         BeanUtils.copyProperties(record, agendamento);
-        MesaEntity mesa = this.mesaRepository.findByNumero(record.mesa().numero()).orElseThrow();
         agendamento.setMesa(mesa);
+        agendamento.setUsuario(user);
 
         this.repository.save(agendamento);
         return new AgendamentoResponseRecord(agendamento);
     }
 
-    public AgendamentoResponseRecord cancelarAgendamento(Long id) {
+    public AgendamentoResponseRecord cancelarAgendamento(Long userId, Long id) {
+        UsuarioEntity user = this.userRepository.findById(userId).orElseThrow(() -> new UsuarioException("Usuário não encontrado", HttpStatus.NOT_FOUND));
+
         AgendamentoEntity agendamento = this.repository.findById(id).orElseThrow(() -> new AgendamentoException("Nenhum agendamento encontrado", HttpStatus.NOT_FOUND));
+
+        if (!Objects.equals(user.getId(), agendamento.getUsuario().getId()))
+            throw new AgendamentoException("Você não tem autorização para mudar essa agendamneto", HttpStatus.UNAUTHORIZED);
 
         this.repository.delete(agendamento);
 
@@ -51,26 +68,40 @@ public class AgendamentoService {
     }
 
     public AgendamentoResponseRecord editarAgendamento(Long id, AgendamentoRecord record) {
+        UsuarioEntity user = this.userRepository.findById(record.userId()).orElseThrow(() -> new UsuarioException("Usuário não encontrado", HttpStatus.NOT_FOUND));
+
         AgendamentoEntity agendamento = this.repository.findById(id)
                 .orElseThrow(() -> new AgendamentoException("Nenhum agendamento encontrado", HttpStatus.NOT_FOUND));
+
+        if (!Objects.equals(user.getId(), agendamento.getUsuario().getId()))
+            throw new AgendamentoException("Você não tem autorização para mudar essa agendamneto", HttpStatus.UNAUTHORIZED);
 
         MesaEntity mesa = this.mesaRepository.findByNumero(record.mesa().numero())
                 .orElseThrow(() -> new AgendamentoException("Mesa informada não encontrada", HttpStatus.NOT_FOUND));
 
-        if (this.repository.existeConflitoNoAgendamento(record.data(), calcHoras(record.data(), record.duracao()), record.mesa().numero(), agendamento.getId()))
+        if (this.repository.existeConflitoNoAgendamento(record.data(), calcHoras(record.data(), record.duracao()), mesa.getId(), agendamento.getId()))
             throw new AgendamentoException("O tempo de agendamento gera conflito com os demais agendamentos", HttpStatus.CONFLICT);
 
         agendamento.setData(record.data());
         agendamento.setDuracao(record.duracao());
         agendamento.setMesa(mesa);
 
-        AgendamentoEntity agendamentoAtualizado = this.repository.save(agendamento);
-
-        return new AgendamentoResponseRecord(agendamentoAtualizado);
+        return new AgendamentoResponseRecord(this.repository.save(agendamento));
     }
 
     public List<AgendamentoResponseRecord> todosAgendamentos() {
         List<AgendamentoEntity> agendamentos = this.repository.findAll();
+
+        if (agendamentos.isEmpty())
+            throw new AgendamentoException("Nenhum agendamento encontrado", HttpStatus.NOT_FOUND);
+
+        return agendamentos.stream().map(AgendamentoResponseRecord::new).toList();
+    }
+
+    public List<AgendamentoResponseRecord> usuarioTodosAgendamentos(Long id) {
+        UsuarioEntity user = this.userRepository.findById(id).orElseThrow(() -> new UsuarioException("Usuário não encontrado", HttpStatus.NOT_FOUND));
+
+        List<AgendamentoEntity> agendamentos = this.repository.findByUsuarioId(user.getId());
 
         if (agendamentos.isEmpty())
             throw new AgendamentoException("Nenhum Agendamento encontrado", HttpStatus.NOT_FOUND);
@@ -78,20 +109,29 @@ public class AgendamentoService {
         return agendamentos.stream().map(AgendamentoResponseRecord::new).toList();
     }
 
-    public AgendamentoResponseRecord agendamentoId(Long id) {
+    public AgendamentoResponseRecord agendamentoId(Long userId, Long id) {
+        UsuarioEntity user = this.userRepository.findById(userId).orElseThrow(() -> new UsuarioException("Usuário não encontrado", HttpStatus.NOT_FOUND));
+
         AgendamentoEntity agendamento = this.repository.findById(id).orElseThrow(() -> new AgendamentoException("Nenhum agendamento encontrado", HttpStatus.NOT_FOUND));
+
+        if (!Objects.equals(user.getId(), agendamento.getUsuario().getId()))
+            throw new AgendamentoException("Você não tem autorização para mudar essa agendamneto", HttpStatus.UNAUTHORIZED);
 
         return new AgendamentoResponseRecord(agendamento);
     }
 
-    public List<AgendamentoResponseRecord> agendamentoDoDia(LocalDate data, Integer numero_mesa) {
+    public List<AgendamentoResponseRecord> agendamentoDoDia(Long userId, LocalDate data, Integer mesaId) {
+        this.userRepository.findById(userId).orElseThrow(() -> new UsuarioException("Usuário não encontrado", HttpStatus.NOT_FOUND));
+
         LocalDateTime inicio = data.atStartOfDay();
         LocalDateTime fim = data.atTime(LocalTime.MAX);
 
-        List<AgendamentoEntity> agendamentos = this.repository.findAgendamentosDoDia(inicio, fim, numero_mesa);
+        MesaEntity mesa = this.mesaRepository.findByNumero(mesaId).orElseThrow(() -> new RestauranteException("Mesa informada não encontrada", HttpStatus.NOT_FOUND));
+
+        List<AgendamentoEntity> agendamentos = this.repository.findAgendamentosDoDia(inicio, fim, mesa.getId(), userId);
 
         if (agendamentos.isEmpty())
-            throw new AgendamentoException("Nenhum agendamento na data " + data + " para a mesa: " + numero_mesa, HttpStatus.NOT_FOUND);
+            throw new AgendamentoException("Nenhum agendamento na data " + data + " para a mesa: " + mesaId, HttpStatus.NOT_FOUND);
 
         return agendamentos.stream().map(AgendamentoResponseRecord::new).toList();
     }
